+++ CHANGELOG.md
@@ -8,6 +8,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Rate limits [#98](https://github.com/rokwire/notifications-building-block/issues/98)
 
+## [1.10.0] - 2022-12-21
+### Added
+- Add message id in the notification [#135](https://github.com/rokwire/notifications-building-block/issues/135)
+- Improve stats API [#133](https://github.com/rokwire/notifications-building-block/issues/133)
+
 ## [1.9.1] - 2022-12-20
 ### Fixed
 - Fix GET messages filter [#131](https://github.com/rokwire/notifications-building-block/issues/131)